+++ dependency_comb/cli/report.py
@@ -42,7 +42,8 @@ from .. import __pkgname__
     metavar="STRING",
     type=click.Choice(AVAILABLE_FORMATS.keys()),
     help="Format name.",
-    default=DEFAULT_FORMAT
+    default=DEFAULT_FORMAT,
+    show_default=True,
 )
 @click.option(
     "--destination",
@@ -59,9 +60,8 @@ from .. import __pkgname__
     type=click.INT,
     default=20,
     help=(
-        "Amount of requirements to process in a chunk. Default to 20 requirements per "
-        "chunk. If zero, it means every requirements are processed in a single "
-        "job without no pause."
+        "Amount of requirements to process in a chunk. If zero, it means every "
+        "requirements are processed in a single job without no pause."
     ),
 )
 @click.option(
@@ -69,7 +69,7 @@ from .. import __pkgname__
     type=click.INT,
     default=1,
     help=(
-        "The time in second to pause before each chunk. Default to 1 second. If zero "
+        "The time in second to pause before each chunk. If zero "
         "it means no pause. Prefer to disable chunk if you don't want any pause."
     ),
 )
@@ -78,8 +78,7 @@ from .. import __pkgname__
     type=click.INT,
     default=15,
     help=(
-        "Timeout in seconds for API requests. The default timeout is set to 15 "
-        "seconds, set it to 0 to disable timeout."
+        "Timeout in seconds for API requests. Set it to 0 to disable timeout."
     ),
 )
 @click.option(
@@ -100,10 +99,12 @@ from .. import __pkgname__
         "environment variable all specifier markers are ignored (so its requirement "
         "is always considered valid)."
     ),
+    show_default=True,
 )
 @click.option(
-    "--failures",
+    "--failures/--no-failures",
     is_flag=True,
+    default=True,
     help=(
         "Include requirement analyze failures in a different table, also each table"
         "will have its own title."
@@ -120,7 +121,7 @@ def report_command(*args, **parameters):
     Arguments:
 
     \b
-    source
+    SOURCE
         Pip requirements file to parse, analyze and report.
         Instead of a file path you can also give a requirements file content from
         standard input using '-'. For example using the Pip freeze output: